Create third layer for DoorView
In order that the doorFrame layer’s frame be transformed
independently of the door itself, a third layer has been added
to Doorview. This third layer acts as the superlayer, of which
the door itself and doorframe are sublayers. Now the door
can be animated to open (i.e. change the frame/bounds)
while leaving the doorframe in its initial position.